Sniper Bot/requirements.txt
python-telegram-bot==20.7
solana==0.29.0
solders==0.18.0
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
base58==2.1.1
